Rename hooks.py to downloaders.py
