+++ pooch/__init__.py
@@ -3,7 +3,7 @@
 from . import version
 from .core import Pooch, create
 from .utils import os_cache, file_hash, make_registry, check_version
-from .hooks import HTTPDownloader
+from .downloaders import HTTPDownloader
 from .processors import Unzip, UnzipSingle
